Francois Tigeot <ftigeot@wolfpond.org>
int -> unsigned long for one sysctl argument, so it works on 64bit FreeBSD.


git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@7262 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ libgphoto2/gphoto2-filesys.c
@@ -1920,7 +1920,7 @@ gp_get_free_memory (GPContext *context, unsigned *free)
 #elif HAVE_SYSCTL && (__FreeBSD__ || __NetBSD__ || __OpenBSD__ || __APPLE__ )
 
 	int mib[2] = { CTL_HW, HW_PHYSMEM };
-	int value;
+	unsigned long value;
 	size_t valuelen = sizeof(value);
 	*free=0;
 	if (sysctl(mib, 2 , &value, &valuelen, NULL, 0) == -1) {
